+++ database/migrations/2022_09_01_104900_create_assets_table.php
@@ -33,7 +33,7 @@ return new class () extends Migration {
             $table->string('code')->index()->collation('und_ns')->comment('mirror of asset model');
             $table->string('name', 255)->nullable()->comment('mirror of asset model');
             $table->decimal('price', 18)->nullable()->comment('mirror of asset model');
-            $table->unsignedSmallInteger('units')->comment('mirror of asset model');
+            $table->decimal('units', 9, 3)->comment('mirror of asset model');
             $table->string('unit')->nullable()->comment('mirror of asset model');
